add transcript button
