+++ src/app/components/public/post-detail/post-detail.component.html
@@ -6,10 +6,6 @@
     <div *ngIf="showAdminUI && postDetail?.isPostNotApproved && loading == false">
         <div *ngIf="postDetail?.isPostRejected == false">
             <!-- when post not rejected -->
-            <div class="fileUpload btn btn-primary pull-left">
-              <span>TRANSCRIPT</span>
-              <input type="file" class="upload" (change)="handleTranscriptButton($event)" name="file" multiple>
-            </div>
 
             <button type="button" class="btn pull-right btn-dangerous"
             (click)="handleRejectButtonClicked()">REJECT</button>
@@ -32,6 +28,11 @@
 
     <button type="button" *ngIf="showAdminUI && !postDetail?.isPostNotApproved && loading === false" class="btn btn-primary pull-right"
       (click)="unapprovePost()">UNAPPROVE</button>
+      <div class="fileUpload pull-left btn btn-primary">
+        <span>ADD TRANSCRIPT</span>
+        <input type="file" class="upload" (change)="handleTranscriptButton($event)" name="file" multiple>
+      </div>
+      
   </div>
 </div>
 <div class="row" style="margin-top:1em;" *ngIf="approvalMessage || unapprovalMessage || assetWarning">
