+++ src/app/submit-form/submit-form.service.ts
@@ -7,8 +7,9 @@ import { NgForm } from "@angular/forms";
 import "rxjs/add/operator/map";
 import "rxjs/add/operator/catch";
 
-import { UploadService } from "./media/upload-service";
+import { UploadService } from "../services/upload-service";
 import { environment } from "../../environments/environment";
+import { UploadLinks } from "app/services/upload-link";
 
 @Injectable()
 export class SubmitFormService {
@@ -40,7 +41,7 @@ export class SubmitFormService {
   filename: string;
   fileInfos: any[] = [];
 
-  constructor(private _http: Http, private _uploadService: UploadService) {
+  constructor(private _http: Http, private _uploadService: UploadService, private uploadLink: UploadLinks) {
     this.title = null;
     this.description = null;
     this.hiddenDescription = null;
@@ -134,7 +135,7 @@ export class SubmitFormService {
         }
         let postId = res.json();
         //console.log(fn + "postId received: " + postId);
-        this.linkFiles(postId);
+        this.uploadLink.linkFiles(postId, this.fileInfos);
         return res;
       })
       .catch((error: any) => {
@@ -143,83 +144,5 @@ export class SubmitFormService {
   }
 
   // TODO: replace assetDescription with form field input from the user.
-  linkFiles(postId: string) {
-    let fn: string = this.constructor.name + "#linkFiles";
-    //console.log(fn + "invoked with postId: " + postId);
-
-    // console.log("fileinfos" + this.fileInfos);
-
-    let jsonLink: any;
-    for (let fileinfo of this.fileInfos) {
-      let string_to_pass = fileinfo.file.name;
-      //console.log(fn + "file key", string_to_pass);
-      jsonLink = {
-        stagingAreaBucketName: this.endPoint.stagingAreaBucketName,
-        assetDescription: "Asset",
-        finalBucketName: this.endPoint.finalBucketName,
-        queueName: this.endPoint.queueName,
-        PostId: postId,
-        key: string_to_pass,
-        tableName: this.endPoint.ddb_table_name
-      };
-
-      if (this.isDocument(string_to_pass)) {
-        this._http
-          .get(this.endPoint.read_file + string_to_pass)
-          .map((res: Response) => res.json())
-          .catch((error: any) => Observable.throw(error.json().error))
-          .subscribe(
-            data => {
-              console.log(data);
-              jsonLink.assetDescription = data;
-            },
-            error => {
-              console.log(error);
-            },
-            () => {
-              this.linkSingleFile(jsonLink);
-            }
-          );
-      } else {
-        this.linkSingleFile(jsonLink);
-      }
-    }
-  }
-
-  private linkSingleFile(jsonLink: any) {
-    //console.log(fn + "data to link", jsonLink);
-    let headers = new Headers();
-    headers.append("Content-Type", "application/json");
-    let options = new RequestOptions({ headers: headers, method: "post" });
-
-    var input = JSON.stringify(jsonLink);
-
-    // returns 504, make admin to check if went through.
-    this._http
-      .post(this.endPoint.link_media, input, options)
-      .map((res: Response) => res.json())
-      .catch((error: any) => Observable.throw(error.json().error))
-      .subscribe(
-        data => {
-          console.log("Link response: ", data);
-        },
-        error => {
-          console.log(error);
-        }
-      );
-  }
-
-  private isDocument(fileName: string): boolean {
-    switch (fileName.substring(fileName.lastIndexOf('.')).toLowerCase()) {
-      case ".doc":
-      case ".docx":
-      case ".rtf":
-      case ".txt":
-      case ".pdf":
-      case ".odt":
-        return true;
-      default:
-        return false;
-    }
-  }
+  
 }
